Refactor naming of attribute(s)Descriptors to singular
