+++ visualization/app/codeCharta/state/selectors/primaryMetrics/primaryMetrics.selector.ts
@@ -4,7 +4,7 @@ import { Metric } from "../../../ui/attributeSideBar/util/metric"
 import { primaryMetricNamesSelector } from "./primaryMetricNames.selector"
 import { createSelector } from "../../angular-redux/createSelector"
 import { getMetricDescriptors } from "../../../ui/attributeSideBar/util/metricDescriptors"
-import { attributeDescriptorsSelector } from "../../store/fileSettings/attributeDescriptors/attributesDescriptors.selector"
+import { attributeDescriptorsSelector } from "../../store/fileSettings/attributeDescriptors/attributeDescriptors.selector"
 import { Edge } from "app/codeCharta/ui/attributeSideBar/util/edge"
 import { CodeMapNode } from "app/codeCharta/codeCharta.model"
 import { AttributeDescriptors } from "../../../codeCharta.model"
