+++ visualization/app/codeCharta/ui/legendPanel/selectors/legendAreaMetric.selector.ts
@@ -1,6 +1,6 @@
 import { createSelector } from "../../../state/angular-redux/createSelector"
 import { areaMetricSelector } from "../../../state/store/dynamicSettings/areaMetric/areaMetric.selector"
-import { attributeDescriptorsSelector } from "../../../state/store/fileSettings/attributeDescriptors/attributesDescriptors.selector"
+import { attributeDescriptorsSelector } from "../../../state/store/fileSettings/attributeDescriptors/attributeDescriptors.selector"
 import { CcState } from "../../../state/store/store"
 import { getMetricDescriptors, MetricDescriptors } from "../../attributeSideBar/util/metricDescriptors"
