+++ visualization/app/codeCharta/ui/legendPanel/selectors/legendColorMetric.selector.ts
@@ -2,7 +2,7 @@ import { createSelector } from "../../../state/angular-redux/createSelector"
 import { colorMetricSelector } from "../../../state/store/dynamicSettings/colorMetric/colorMetric.selector"
 import { CcState } from "../../../state/store/store"
 import { getMetricDescriptors, MetricDescriptors } from "../../attributeSideBar/util/metricDescriptors"
-import { attributeDescriptorsSelector } from "../../../state/store/fileSettings/attributeDescriptors/attributesDescriptors.selector"
+import { attributeDescriptorsSelector } from "../../../state/store/fileSettings/attributeDescriptors/attributeDescriptors.selector"
 
 export const legendColorMetricSelector: (state: CcState) => MetricDescriptors = createSelector(
 	[colorMetricSelector, attributeDescriptorsSelector],
