+++ visualization/app/codeCharta/ui/legendPanel/selectors/legendHeightMetric.selector.ts
@@ -2,7 +2,7 @@ import { createSelector } from "../../../state/angular-redux/createSelector"
 import { heightMetricSelector } from "../../../state/store/dynamicSettings/heightMetric/heightMetric.selector"
 import { CcState } from "../../../state/store/store"
 import { getMetricDescriptors, MetricDescriptors } from "../../attributeSideBar/util/metricDescriptors"
-import { attributeDescriptorsSelector } from "../../../state/store/fileSettings/attributeDescriptors/attributesDescriptors.selector"
+import { attributeDescriptorsSelector } from "../../../state/store/fileSettings/attributeDescriptors/attributeDescriptors.selector"
 
 export const legendHeightMetricSelector: (state: CcState) => MetricDescriptors = createSelector(
 	[heightMetricSelector, attributeDescriptorsSelector],
